requirements.txt
alembic==1.10.4
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.3.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==1.5
requests==2.30.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.0
WTForms==3.0.1
